fix(mobile): resolve iOS SwiftLint and Android compile CI failures
Move fetchQuizProctoringConfig into LMSAPIProctoring.swift so
LMSAPIFeatures.swift stays within the SwiftLint file_length limit.
Use ActivityManager.lockTaskModeState instead of the unresolved
Activity.isInLockTaskMode Kotlin binding.
